Get helpers out of matrix_box_presenter, divvy up the thumbnail helpers a bit
