+++ app/helpers/image_votes_helper.rb
@@ -1,8 +1,35 @@
 # frozen_string_literal: true
 
-# image vote lookup used in show_image
 module ImageVotesHelper
-  # moved from the ImagesController
+  # used in shared/image_thumbnail
+  def vote_section_html(votes, image)
+    return "" unless votes && image && User.current
+
+    content_tag(:div, "", class: "vote-section") do
+      render(partial: "shared/image_vote_links",
+             locals: { image: image })
+    end
+  end
+
+  # Create an image link vote, where vote param is vote number ie: 3
+  # Returns a form input button if the user has NOT voted this way
+  # JS is listening to any element with [data-role="image_vote"],
+  # Even though this is not an <a> tag, but an <input>, it's ok.
+  def image_vote_link(image, vote)
+    current_vote = image.users_vote(@user)
+    vote_text = vote.zero? ? "(x)" : image_vote_as_short_string(vote)
+
+    if current_vote == vote
+      return content_tag(:span, image_vote_as_short_string(vote))
+    end
+
+    put_button(name: vote_text, remote: true,
+               path: image_vote_path(image_id: image.id, value: vote),
+               title: image_vote_as_help_string(vote),
+               data: { role: "image_vote", image_id: image.id, value: vote })
+  end
+
+  # image vote lookup used in show_image
   def find_list_of_votes(image)
     image.image_votes.sort_by do |v|
       (v.anonymous ? :anonymous.l : v.user.unique_text_name).downcase
