+++ app/views/shared/_matrix_box.html.erb
@@ -19,25 +19,18 @@ if presenter
           end %>
         </div><!-- .thumbnail-container -->
         <div class="rss-what pt-2">
-          <h5><%= presenter.what %></h5>
-          <%= if identify
-            if (object.name_id != 1) && (nam = object.consensus_naming)
-              content_tag(:div, class: "vote-select-container mb-3") do
-                render(partial: "observations/namings/votes/form",
-                       locals: { naming: nam })
-              end
-            else
-              propose_naming_link(object_id, "btn-default mb-3")
-            end
-          end %>
+          <h5><%= link_with_query(presenter.name,
+                                  presenter.what.show_link_args) %></h5>
+          <%= vote_or_propose_ui(identify, object) %>
         </div><!-- .rss-what -->
         <div class="rss-where">
-          <small><%= presenter.where %></small>
+          <small><%= location_link(presenter.place_name,
+                                   presenter.where) %></small>
         </div><!-- .rss-where -->
         <div class="rss-what">
           <small class="nowrap-ellipsis">
             <% unless presenter.when.blank? %>
-              <%= presenter.when %>: <%= presenter.who %>
+              <%= presenter.when %>: <%= user_link(presenter.who) %>
             <% end %>
           </small>
         </div><!-- .rss-when -->
@@ -48,14 +41,7 @@ if presenter
           <small><%= presenter.fancy_time %></small>
         </div><!-- .rss-fancy-time -->
       </div><!-- .panel-body -->
-      <% if identify %>
-        <%= content_tag(
-              :div,
-              class: "panel-footer panel-active text-center position-relative"
-            ) do
-          mark_as_reviewed_toggle(object_id, "box_reviewed", "stretched-link")
-        end %>
-      <% end %>
+      <%= identify_footer(identify, object_id) %>
     </div><!-- .panel -->
   </li><!-- .matrix-box -->
 <% end %>
